requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tqdm>=4.60.0
Pillow>=8.2.0
seaborn>=0.11.0
onnx>=1.10.0
onnxruntime>=1.8.0
opencv-python>=4.5.0
pathlib>=1.0.1